Remove legacy scheduler artifacts
