+++ src/core/task_management/README.md
@@ -25,28 +25,26 @@ This directory has been consolidated from 6 fragmented task scheduler files into
 ```
 src/core/task_management/
 ├── __init__.py                    # Unified package exports
-├── unified_task_scheduler.py      # Consolidated task scheduler (ALL functionality)
+├── unified_scheduler/             # Modular scheduler components
+│   ├── __init__.py               # Scheduler package interface
+│   ├── enums.py                  # Task enums
+│   ├── models.py                 # Task data models
+│   ├── metrics.py                # Scheduling metrics
+│   └── scheduler/                # Scheduler implementation mixins
 └── README.md                     # This documentation
 ```
 
 ### **Key Components:**
 
-#### **1. unified_task_scheduler.py** - Complete Task Management System
-- **Class**: `UnifiedTaskScheduler`
-- **Responsibility**: Complete task scheduling and management
-- **Features**: 
-  - Task creation, validation, and submission
-  - Priority-based scheduling and assignment
-  - Dependency resolution and validation
-  - Resource allocation and tracking
-  - Metrics collection and monitoring
-  - Task coordination and execution
-  - Event callbacks and monitoring
-  - Smoke testing and validation
+#### **1. unified_scheduler/** - Modular Scheduler Subpackage
+- **enums.py**: Task definitions and enums
+- **models.py**: Core task data structures
+- **metrics.py**: Scheduling metrics and performance tracking
+- **scheduler/**: Mixins composing the `UnifiedTaskScheduler`
 
 #### **2. __init__.py** - Unified Package Interface
 - **Purpose**: Single entry point for all task management functionality
-- **Imports**: Consolidated from unified_task_scheduler.py
+- **Imports**: Consolidated from unified_scheduler modules
 - **Exports**: All task management classes, types, and functions
 
 ---
@@ -85,7 +83,7 @@ from src.core.task_management import (
 ### **Total Lines Eliminated: 941 lines of fragmented code**
 
 ### **Files Created:**
-- ✅ `unified_task_scheduler.py` (600+ lines) - All functionality consolidated
+- ✅ `unified_scheduler/` subpackage (scheduler mixins, enums.py, models.py, metrics.py) - Modular task management components
 - ✅ `__init__.py` - Updated to use unified system
 - ✅ `README.md` - Comprehensive documentation
 
@@ -182,4 +180,3 @@ print(f"Smoke test: {'PASSED' if success else 'FAILED'}")
 **The task scheduler system has been successfully consolidated from 6 fragmented files into a unified, V2-compliant system. All functionality is preserved, fragmentation is eliminated, and the architecture follows V2 standards perfectly.**
 
 **WE. ARE. SWARM. - Consolidation complete! 🚀**
-
