+++ src/core/task_management/unified_scheduler/__init__.py
@@ -1,0 +1,35 @@
+from .enums import TaskPriority, TaskStatus, TaskType, TaskCategory
+from .models import (
+    Task,
+    TaskDependency,
+    TaskResource,
+    TaskConstraint,
+    TaskMetadata,
+)
+from .metrics import SchedulingMetrics
+from .scheduler import UnifiedTaskScheduler
+
+TaskScheduler = UnifiedTaskScheduler
+TaskSchedulerConfig = UnifiedTaskScheduler
+TaskSchedulerManager = UnifiedTaskScheduler
+TaskSchedulerCoordinator = UnifiedTaskScheduler
+TaskSchedulerCore = UnifiedTaskScheduler
+
+__all__ = [
+    "UnifiedTaskScheduler",
+    "TaskScheduler",
+    "Task",
+    "TaskPriority",
+    "TaskStatus",
+    "TaskType",
+    "TaskCategory",
+    "TaskDependency",
+    "TaskResource",
+    "TaskConstraint",
+    "TaskMetadata",
+    "SchedulingMetrics",
+    "TaskSchedulerConfig",
+    "TaskSchedulerManager",
+    "TaskSchedulerCoordinator",
+    "TaskSchedulerCore",
+]
