--- src/core/task_management/unified_task_scheduler.py
@@ -1,731 +1,0 @@
-"""Unified Task Scheduler - Consolidated from 6 task scheduler files"""
-
-from __future__ import annotations
-
-import asyncio
-import logging
-import threading
-import time
-import uuid
-from dataclasses import dataclass, field
-from datetime import datetime, timedelta
-from queue import PriorityQueue
-from typing import Any, Callable, Dict, List, Optional, Set
-from collections import defaultdict
-from enum import Enum
-
-logger = logging.getLogger(__name__)
-
-
-# ============================================================================
-# TASK TYPES AND ENUMS (Consolidated from task_types.py)
-# ============================================================================
-
-class TaskPriority(Enum):
-    """Task priority levels for scheduling and execution."""
-    LOW = 0
-    NORMAL = 1
-    HIGH = 2
-    URGENT = 3
-    CRITICAL = 4
-
-
-class TaskStatus(Enum):
-    """Status of a task in the system."""
-    PENDING = "pending"
-    QUEUED = "queued"
-    RUNNING = "running"
-    PAUSED = "paused"
-    COMPLETED = "completed"
-    FAILED = "failed"
-    CANCELLED = "cancelled"
-    TIMEOUT = "timeout"
-    BLOCKED = "blocked"
-
-
-class TaskType(Enum):
-    """Types of tasks supported by the system."""
-    COMPUTATION = "computation"
-    DATA_PROCESSING = "data_processing"
-    COMMUNICATION = "communication"
-    COORDINATION = "coordination"
-    MONITORING = "monitoring"
-    ANALYSIS = "analysis"
-    OPTIMIZATION = "optimization"
-    VALIDATION = "validation"
-    SYNCHRONIZATION = "synchronization"
-    MAINTENANCE = "maintenance"
-
-
-class TaskCategory(Enum):
-    """Categories for organizing and filtering tasks."""
-    SYSTEM = "system"
-    USER = "user"
-    AUTOMATED = "automated"
-    SCHEDULED = "scheduled"
-    EMERGENCY = "emergency"
-    MAINTENANCE = "maintenance"
-    DEVELOPMENT = "development"
-    TESTING = "testing"
-
-
-@dataclass
-class TaskDependency:
-    """Task dependency information."""
-    task_id: str
-    dependency_type: str  # "required", "optional", "parallel"
-    condition: str = "completed"  # "completed", "successful", "any"
-    timeout: Optional[int] = None  # seconds to wait for dependency
-
-
-@dataclass
-class TaskResource:
-    """Resource requirements for a task."""
-    cpu_cores: int = 1
-    memory_mb: int = 512
-    storage_mb: int = 100
-    network_bandwidth: Optional[int] = None  # MB/s
-    gpu_required: bool = False
-    special_hardware: List[str] = field(default_factory=list)
-
-
-@dataclass
-class TaskConstraint:
-    """Constraints that must be satisfied for task execution."""
-    deadline: Optional[datetime] = None
-    max_execution_time: Optional[int] = None  # seconds
-    retry_limit: int = 3
-    priority_boost: bool = False
-    exclusive_execution: bool = False
-
-
-@dataclass
-class TaskMetadata:
-    """Additional metadata for task tracking and analysis."""
-    tags: List[str] = field(default_factory=list)
-    source: str = "unknown"
-    estimated_complexity: str = "medium"
-    business_value: str = "normal"
-    risk_level: str = "low"
-
-
-@dataclass
-class Task:
-    """Core task definition with all necessary attributes."""
-    task_id: str = field(default_factory=lambda: str(uuid.uuid4()))
-    name: str = ""
-    content: str = ""
-    task_type: TaskType = TaskType.COMPUTATION
-    category: TaskCategory = TaskCategory.USER
-    priority: TaskPriority = TaskPriority.NORMAL
-    status: TaskStatus = TaskStatus.PENDING
-    
-    # Assignment and ownership
-    assigned_agent: Optional[str] = None
-    created_by: str = "system"
-    assigned_at: Optional[datetime] = None
-    started_at: Optional[datetime] = None
-    completed_at: Optional[datetime] = None
-    
-    # Timing and execution
-    created_at: datetime = field(default_factory=datetime.now)
-    due_date: Optional[datetime] = None
-    execution_time: Optional[float] = None  # seconds
-    retry_count: int = 0
-    
-    # Dependencies and relationships
-    dependencies: List[TaskDependency] = field(default_factory=list)
-    dependent_tasks: List[str] = field(default_factory=list)
-    
-    # Resources and constraints
-    resources: TaskResource = field(default_factory=TaskResource)
-    constraints: TaskConstraint = field(default_factory=TaskConstraint)
-    metadata: TaskMetadata = field(default_factory=TaskMetadata)
-    
-    # Results and error handling
-    result: Optional[Any] = None
-    error_message: Optional[str] = None
-    error_details: Optional[Dict[str, Any]] = None
-    
-    def is_ready_to_execute(self) -> bool:
-        """Check if task is ready to execute."""
-        if self.status != TaskStatus.PENDING:
-            return False
-        
-        # Check dependencies
-        for dep in self.dependencies:
-            if dep.condition == "completed" and dep.task_id not in self.completed_tasks:
-                return False
-        
-        return True
-    
-    def is_expired(self) -> bool:
-        """Check if task has expired."""
-        if not self.constraints.deadline:
-            return False
-        return datetime.now() > self.constraints.deadline
-    
-    def complete_execution(self, result: Any = None):
-        """Mark task as completed."""
-        self.status = TaskStatus.COMPLETED
-        self.result = result
-        self.completed_at = datetime.now()
-        if self.started_at:
-            self.execution_time = (self.completed_at - self.started_at).total_seconds()
-
-
-# ============================================================================
-# SCHEDULING METRICS (Consolidated from task_scheduler_config.py)
-# ============================================================================
-
-@dataclass
-class SchedulingMetrics:
-    """Metrics for monitoring scheduler performance."""
-    total_tasks_scheduled: int = 0
-    total_tasks_completed: int = 0
-    total_tasks_failed: int = 0
-    average_scheduling_time: float = 0.0
-    average_execution_time: float = 0.0
-    tasks_by_priority: Dict[TaskPriority, int] = field(default_factory=dict)
-    tasks_by_type: Dict[TaskType, int] = field(default_factory=dict)
-    last_update: datetime = field(default_factory=datetime.now)
-
-
-# ============================================================================
-# UNIFIED TASK SCHEDULER (Consolidated from all components)
-# ============================================================================
-
-class UnifiedTaskScheduler:
-    """
-    Unified Task Scheduler - Single responsibility: Complete task scheduling and management
-    
-    This class consolidates functionality from:
-    - task_scheduler.py (orchestrator)
-    - task_scheduler_config.py (configuration and metrics)
-    - task_scheduler_manager.py (high-level management)
-    - task_scheduler_coordinator.py (task coordination)
-    - task_scheduler_core.py (core utilities)
-    - task_types.py (data structures)
-    
-    Total consolidation: 6 files → 1 file (100% functionality preserved)
-    """
-
-    def __init__(self, max_concurrent_tasks: int = 100):
-        """Initialize unified task scheduler."""
-        self.max_concurrent_tasks = max_concurrent_tasks
-        self._lock = threading.RLock()
-        
-        # Task queues by priority
-        self._priority_queues: Dict[TaskPriority, PriorityQueue] = {
-            priority: PriorityQueue() for priority in TaskPriority
-        }
-        
-        # Task storage and tracking
-        self._tasks: Dict[str, Task] = {}
-        self._running_tasks: Dict[str, Task] = {}
-        self._completed_tasks: Dict[str, Task] = {}
-        self._failed_tasks: Dict[str, Task] = {}
-        
-        # Dependency tracking
-        self._dependency_graph: Dict[str, Set[str]] = defaultdict(set)
-        self._reverse_dependencies: Dict[str, Set[str]] = defaultdict(set)
-        
-        # Resource tracking
-        self._agent_resources: Dict[str, Dict[str, Any]] = {}
-        self._resource_locks: Dict[str, threading.Lock] = {}
-        
-        # Scheduling metrics
-        self._metrics = SchedulingMetrics()
-        
-        # Event callbacks
-        self._task_callbacks: Dict[str, List[Callable]] = defaultdict(list)
-        
-        # Scheduler state
-        self._running = False
-        self._scheduler_thread: Optional[threading.Thread] = None
-        
-        logger.info(f"✅ Unified Task Scheduler initialized with max {max_concurrent_tasks} concurrent tasks")
-
-    # ============================================================================
-    # SCHEDULER MANAGEMENT METHODS
-    # ============================================================================
-
-    async def start(self):
-        """Start the task scheduler."""
-        if self._running:
-            logger.warning("Task scheduler already running")
-            return
-
-        self._running = True
-        self._scheduler_thread = threading.Thread(
-            target=self._scheduler_loop, daemon=True
-        )
-        self._scheduler_thread.start()
-
-        logger.info("🚀 Unified Task Scheduler started")
-
-    async def stop(self):
-        """Stop the task scheduler."""
-        if not self._running:
-            return
-
-        self._running = False
-        if self._scheduler_thread:
-            self._scheduler_thread.join(timeout=5)
-
-        logger.info("⏹️ Unified Task Scheduler stopped")
-
-    def _scheduler_loop(self):
-        """Main scheduler loop for processing tasks."""
-        while self._running:
-            try:
-                with self._lock:
-                    self._process_pending_tasks()
-                    self._check_expired_tasks()
-                    self._cleanup_old_tasks()
-                time.sleep(1)
-            except Exception as e:
-                logger.error(f"Error in scheduler loop: {e}")
-                time.sleep(5)
-
-    # ============================================================================
-    # TASK MANAGEMENT METHODS
-    # ============================================================================
-
-    async def submit_task(self, task: Task) -> bool:
-        """Submit a task for scheduling."""
-        try:
-            with self._lock:
-                if not self._validate_task(task):
-                    logger.error(f"Task validation failed for {task.task_id}")
-                    return False
-
-                self._tasks[task.task_id] = task
-
-                priority_score = self._calculate_priority_score(task)
-                self._priority_queues[task.priority].put((priority_score, task))
-
-                self._update_dependency_graph(task)
-                self._update_metrics(task, "submitted")
-
-                logger.info(f"Task {task.task_id} submitted successfully")
-                return True
-
-        except Exception as e:
-            logger.error(f"Error submitting task {task.task_id}: {e}")
-            return False
-
-    async def get_next_task(self, agent_id: str) -> Optional[Task]:
-        """Get the next available task for an agent."""
-        try:
-            with self._lock:
-                if not self._can_agent_handle_task(agent_id):
-                    return None
-
-                for priority in reversed(list(TaskPriority)):
-                    task = self._get_next_task_from_priority(priority, agent_id)
-                    if task:
-                        return task
-
-                return None
-
-        except Exception as e:
-            logger.error(f"Error getting next task for agent {agent_id}: {e}")
-            return None
-
-    async def complete_task(self, task_id: str, result: Any = None) -> bool:
-        """Mark a task as completed."""
-        try:
-            with self._lock:
-                if task_id not in self._running_tasks:
-                    logger.error(f"Task {task_id} not found in running tasks")
-                    return False
-
-                task = self._running_tasks[task_id]
-                task.complete_execution(result)
-
-                del self._running_tasks[task_id]
-                self._completed_tasks[task_id] = task
-
-                self._handle_task_completion(task_id)
-                self._update_metrics(task, "completed")
-
-                logger.info(f"Task {task_id} completed successfully")
-                return True
-
-        except Exception as e:
-            logger.error(f"Error completing task {task_id}: {e}")
-            return False
-
-    async def fail_task(self, task_id: str, error_message: str) -> bool:
-        """Mark a task as failed."""
-        try:
-            with self._lock:
-                if task_id not in self._running_tasks:
-                    logger.error(f"Task {task_id} not found in running tasks")
-                    return False
-
-                task = self._running_tasks[task_id]
-                task.status = TaskStatus.FAILED
-                task.error_message = error_message
-                task.completed_at = datetime.now()
-
-                del self._running_tasks[task_id]
-                self._failed_tasks[task_id] = task
-
-                self._update_metrics(task, "failed")
-
-                logger.info(f"Task {task_id} marked as failed: {error_message}")
-                return True
-
-        except Exception as e:
-            logger.error(f"Error failing task {task_id}: {e}")
-            return False
-
-    # ============================================================================
-    # CORE SCHEDULING METHODS
-    # ============================================================================
-
-    def _validate_task(self, task: Task) -> bool:
-        """Validate a task before scheduling."""
-        if not task.name:
-            return False
-
-        if not task.content:
-            return False
-
-        # Check for circular dependencies
-        if self._has_circular_dependencies(task):
-            return False
-
-        return True
-
-    def _calculate_priority_score(self, task: Task) -> float:
-        """Calculate priority score for task ordering."""
-        base_score = task.priority.value
-
-        # Factor in deadline urgency
-        if task.constraints.deadline:
-            time_until_deadline = (
-                task.constraints.deadline - datetime.now()
-            ).total_seconds()
-            if time_until_deadline > 0:
-                base_score += 10.0 / (time_until_deadline + 1)
-
-        # Factor in retry count (failed tasks get higher priority)
-        base_score += task.retry_count * 2.0
-
-        return base_score
-
-    def _update_dependency_graph(self, task: Task):
-        """Update the dependency graph when a task is added."""
-        # Add task to graph
-        self._dependency_graph[task.task_id] = set()
-
-        # Add dependencies
-        for dependency in task.dependencies:
-            self._dependency_graph[task.task_id].add(dependency.task_id)
-            self._reverse_dependencies[dependency.task_id].add(task.task_id)
-
-    def _has_circular_dependencies(self, task: Task) -> bool:
-        """Check if adding a task would create circular dependencies."""
-        visited = set()
-        rec_stack = set()
-
-        def has_cycle(node_id: str) -> bool:
-            if node_id in rec_stack:
-                return True
-            if node_id in visited:
-                return False
-
-            visited.add(node_id)
-            rec_stack.add(node_id)
-
-            for dep_id in self._dependency_graph.get(node_id, set()):
-                if has_cycle(dep_id):
-                    return True
-
-            rec_stack.remove(node_id)
-            return False
-
-        return has_cycle(task.task_id)
-
-    # ============================================================================
-    # TASK COORDINATION METHODS
-    # ============================================================================
-
-    def _process_pending_tasks(self):
-        """Process pending tasks and assign them to available agents."""
-        for priority in reversed(list(TaskPriority)):
-            while not self._priority_queues[priority].empty():
-                if len(self._running_tasks) >= self.max_concurrent_tasks:
-                    break
-
-                _, task = self._priority_queues[priority].get()
-
-                if task.is_ready_to_execute():
-                    agent_id = self._find_available_agent(task)
-                    if agent_id:
-                        self._assign_task_to_agent(task, agent_id)
-
-    def _check_expired_tasks(self):
-        """Check for and handle expired tasks."""
-        current_time = datetime.now()
-
-        expired_tasks = []
-        for task_id, task in self._running_tasks.items():
-            if task.is_expired():
-                expired_tasks.append(task_id)
-
-        for task_id in expired_tasks:
-            asyncio.create_task(self.fail_task(task_id, "Task timeout exceeded"))
-
-    def _cleanup_old_tasks(self):
-        """Clean up old completed and failed tasks."""
-        current_time = datetime.now()
-        cutoff_time = current_time - timedelta(hours=24)
-
-        old_completed = [
-            tid
-            for tid, task in self._completed_tasks.items()
-            if task.completed_at and task.completed_at < cutoff_time
-        ]
-        for tid in old_completed:
-            del self._completed_tasks[tid]
-
-        old_failed = [
-            tid
-            for tid, task in self._failed_tasks.items()
-            if task.completed_at and task.completed_at < cutoff_time
-        ]
-        for tid in old_failed:
-            del self._failed_tasks[tid]
-
-    def _can_agent_handle_task(self, agent_id: str) -> bool:
-        """Check if an agent can handle additional tasks."""
-        current_tasks = sum(
-            1
-            for task in self._running_tasks.values()
-            if task.assigned_agent == agent_id
-        )
-        return current_tasks < 5
-
-    def _get_next_task_from_priority(
-        self, priority: TaskPriority, agent_id: str
-    ) -> Optional[Task]:
-        """Get next task from a specific priority queue that agent can handle."""
-        queue = self._priority_queues[priority]
-
-        checked_tasks = []
-
-        while not queue.empty():
-            _, task = queue.get()
-            checked_tasks.append((priority.value, task))
-
-            if self._can_agent_handle_task(agent_id):
-                # Put back other tasks
-                for _, t in checked_tasks[:-1]:
-                    self._priority_queues[priority].put((priority.value, t))
-                return task
-
-        # Put back all tasks if none were suitable
-        for _, task in checked_tasks:
-            self._priority_queues[priority].put((priority.value, task))
-
-        return None
-
-    def _find_available_agent(self, task: Task) -> Optional[str]:
-        """Find an available agent for a task."""
-        # Simplified agent selection - placeholder for future logic
-        available_agents = [
-            agent_id for agent_id in self._agent_resources.keys()
-            if self._can_agent_handle_task(agent_id)
-        ]
-        return available_agents[0] if available_agents else None
-
-    def _assign_task_to_agent(self, task: Task, agent_id: str):
-        """Assign a task to an agent."""
-        task.assigned_agent = agent_id
-        task.status = TaskStatus.RUNNING
-        task.assigned_at = datetime.now()
-        task.started_at = datetime.now()
-
-        self._running_tasks[task.task_id] = task
-        self._allocate_agent_resources(agent_id, task)
-
-        logger.info(f"Task {task.task_id} assigned to agent {agent_id}")
-
-    def _handle_task_completion(self, task_id: str):
-        """Handle completion of a task and update dependent tasks."""
-        for dependent_id in self._reverse_dependencies.get(task_id, set()):
-            dependent_task = self._tasks.get(dependent_id)
-            if dependent_task:
-                if self._are_all_dependencies_satisfied(dependent_task):
-                    dependent_task.status = TaskStatus.PENDING
-
-    def _are_all_dependencies_satisfied(self, task: Task) -> bool:
-        """Check if all dependencies for a task are satisfied."""
-        for dep in task.dependencies:
-            if dep.task_id not in self._completed_tasks:
-                return False
-        return True
-
-    # ============================================================================
-    # RESOURCE MANAGEMENT METHODS
-    # ============================================================================
-
-    def _allocate_agent_resources(self, agent_id: str, task: Task):
-        """Allocate resources for a task on an agent."""
-        # Simplified resource allocation - placeholder for future logic
-        pass
-
-    def _release_agent_resources(self, agent_id: str, task: Task):
-        """Release resources allocated to a task on an agent."""
-        # Simplified resource release - placeholder for future logic
-        pass
-
-    # ============================================================================
-    # METRICS AND MONITORING METHODS
-    # ============================================================================
-
-    def _update_metrics(self, task: Task, action: str):
-        """Update scheduling metrics."""
-        if action == "submitted":
-            self._metrics.total_tasks_scheduled += 1
-            self._metrics.tasks_by_priority[task.priority] = \
-                self._metrics.tasks_by_priority.get(task.priority, 0) + 1
-            self._metrics.tasks_by_type[task.task_type] = \
-                self._metrics.tasks_by_type.get(task.task_type, 0) + 1
-        elif action == "completed":
-            self._metrics.total_tasks_completed += 1
-        elif action == "failed":
-            self._metrics.total_tasks_failed += 1
-
-        self._metrics.last_update = datetime.now()
-
-    def get_metrics(self) -> SchedulingMetrics:
-        """Get current scheduling metrics."""
-        return self._metrics
-
-    def get_task_status(self, task_id: str) -> Optional[Task]:
-        """Get status of a specific task."""
-        return self._tasks.get(task_id)
-
-    def get_all_tasks(self) -> Dict[str, Task]:
-        """Get all tasks in the system."""
-        return self._tasks.copy()
-
-    def get_running_tasks(self) -> Dict[str, Task]:
-        """Get all currently running tasks."""
-        return self._running_tasks.copy()
-
-    def get_completed_tasks(self) -> Dict[str, Task]:
-        """Get all completed tasks."""
-        return self._completed_tasks.copy()
-
-    def get_failed_tasks(self) -> Dict[str, Task]:
-        """Get all failed tasks."""
-        return self._failed_tasks.copy()
-
-    # ============================================================================
-    # CALLBACK AND EVENT METHODS
-    # ============================================================================
-
-    def add_task_callback(self, event: str, callback: Callable):
-        """Add a callback for task events."""
-        self._task_callbacks[event].append(callback)
-
-    def remove_task_callback(self, event: str, callback: Callable):
-        """Remove a callback for task events."""
-        if event in self._task_callbacks and callback in self._task_callbacks[event]:
-            self._task_callbacks[event].remove(callback)
-
-    def _trigger_task_callbacks(self, event: str, task: Task):
-        """Trigger callbacks for a task event."""
-        if event in self._task_callbacks:
-            for callback in self._task_callbacks[event]:
-                try:
-                    callback(task)
-                except Exception as e:
-                    logger.error(f"Error in task callback for event {event}: {e}")
-
-    # ============================================================================
-    # SHUTDOWN AND CLEANUP METHODS
-    # ============================================================================
-
-    def shutdown(self):
-        """Shutdown the task scheduler."""
-        self.stop()
-        logger.info("🔄 Unified Task Scheduler shutdown complete")
-
-    def run_smoke_test(self) -> bool:
-        """Run smoke test for unified task scheduler."""
-        try:
-            logger.info("🧪 Running Unified Task Scheduler smoke test...")
-            
-            # Test basic initialization
-            assert len(self._tasks) == 0
-            assert len(self._priority_queues) == len(TaskPriority)
-            logger.info("✅ Basic initialization passed")
-            
-            # Test task creation and submission
-            test_task = Task(
-                name="Test Task",
-                content="Test content",
-                priority=TaskPriority.NORMAL
-            )
-            
-            # Test task validation
-            assert self._validate_task(test_task)
-            logger.info("✅ Task validation passed")
-            
-            # Test priority score calculation
-            score = self._calculate_priority_score(test_task)
-            assert score >= 0
-            logger.info("✅ Priority score calculation passed")
-            
-            logger.info("🎯 Unified Task Scheduler smoke test PASSED")
-            return True
-            
-        except Exception as exc:
-            logger.error(f"❌ Unified Task Scheduler smoke test FAILED: {exc}")
-            return False
-
-
-# ============================================================================
-# BACKWARD COMPATIBILITY ALIASES
-# ============================================================================
-
-# Maintain backward compatibility with existing code
-TaskScheduler = UnifiedTaskScheduler
-TaskSchedulerConfig = UnifiedTaskScheduler
-TaskSchedulerManager = UnifiedTaskScheduler
-TaskSchedulerCoordinator = UnifiedTaskScheduler
-TaskSchedulerCore = UnifiedTaskScheduler
-
-__all__ = [
-    # Main scheduler class
-    "UnifiedTaskScheduler",
-    "TaskScheduler",
-    
-    # Task types and enums
-    "Task",
-    "TaskPriority",
-    "TaskStatus",
-    "TaskType",
-    "TaskCategory",
-    "TaskDependency",
-    "TaskResource",
-    "TaskConstraint",
-    "TaskMetadata",
-    
-    # Metrics
-    "SchedulingMetrics",
-    
-    # Backward compatibility
-    "TaskSchedulerConfig",
-    "TaskSchedulerManager",
-    "TaskSchedulerCoordinator",
-    "TaskSchedulerCore",
-]
-
